# numba optimizations for gaussians