@bd-j 2019-03-19
Awesome, thanks @lgarrison!!  I'll take a closer look but the changes to the interface and data model seem minor and acceptable; though I don't what making `ImageGaussian` a jitclass might affect.  The factor ~5 speedup in `get_gaussian_gradients` is very nice indeed.  I'm also pleasantly surprised that `compute_gaussian` could be sped up so much.

My experience with Scipy interpolation methods is that with care they can be sped up quite a bit, but I haven't looked at BivariateSpline in particular.

I'll add a `requirements.txt` for the numba version info.